Fix cyclic import bug
